fix(#690): move pagination, mutation_hooks, adaptive_ttl from core/ to lib/
Three more zero-nexus-dependency utility modules extracted from core/:
- pagination.py: cursor-based keyset pagination (stdlib only)
- mutation_hooks.py: MutationOp/MutationEvent/PostMutationHook (stdlib only)
- adaptive_ttl.py: write-frequency-based cache TTL mixin (stdlib only)

All are pure data structures or algorithms with no kernel state coupling.
