+++ src/nexus/services/search_listing_mixin.py
@@ -973,7 +973,7 @@ class SearchListingMixin:
     ) -> Any:
         """Paginated list with over-fetch strategy for permission filtering (Issue #937)."""
         from nexus.core.metadata import PaginatedResult
-        from nexus.core.pagination import encode_cursor
+        from nexus.lib.pagination import encode_cursor
 
         context = context or self._default_context
         import time as _time
