+++ tests/unit/services/test_workflow_dispatch.py
@@ -13,8 +13,8 @@ from unittest.mock import AsyncMock, patch
 import pytest
 
 from nexus.core.metadata import FileMetadata
-from nexus.core.mutation_hooks import MutationEvent, MutationOp
 from nexus.core.pipe_manager import PipeManager
+from nexus.lib.mutation_hooks import MutationEvent, MutationOp
 from nexus.services.workflow_dispatch_service import WorkflowDispatchService
 
 # ======================================================================
